+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - New #93: Add `paidMedia` parameter to `Message` and `ExternalReplyInfo` types, add `PaidMediaInfo`,
   `PaidMediaPreview`, `PaidMediaPhoto`, `PaidMediaVideo` types.
 - New #94: Add `invoicePayload` parameter to `TransactionPartnerUser` type.
-- New #95: Add `InputPaidMediaPhoto`, `InputPaidMediaVideo` types.
+- New #95: Add `sendPaidMedia` method and `InputPaidMediaPhoto`, `InputPaidMediaVideo` types.
 
 ## 0.2.0 June 29, 2024